Rename metadata to metaTokens
